+++ test/bulk-edit/e2e.spec.ts
@@ -4,7 +4,7 @@ import type { PayloadTestSDK } from 'helpers/sdk/index.js'
 import { expect, test } from '@playwright/test'
 import { addArrayRow } from 'helpers/e2e/fields/array/index.js'
 import { selectInput } from 'helpers/e2e/selectInput.js'
-import { addListFilter } from 'helpers/e2e/tables/filters/addListFilter.js'
+import { addListFilter } from 'helpers/e2e/filters/addListFilter.js'
 import { toggleBlockOrArrayRow } from 'helpers/e2e/toggleCollapsible.js'
 import * as path from 'path'
 import { wait } from 'payload/shared'
