+++ test/fields/collections/Relationship/e2e.spec.ts
@@ -4,7 +4,7 @@ import { expect, test } from '@playwright/test'
 import { openCreateDocDrawer } from 'helpers/e2e/fields/relationship/openCreateDocDrawer.js'
 import { navigateToDoc } from 'helpers/e2e/navigateToDoc.js'
 import { openDocControls } from 'helpers/e2e/openDocControls.js'
-import { addListFilter } from 'helpers/e2e/tables/filters/addListFilter.js'
+import { addListFilter } from 'helpers/e2e/filters/addListFilter.js'
 import { openDocDrawer } from 'helpers/e2e/toggleDocDrawer.js'
 import path from 'path'
 import { wait } from 'payload/shared'
