+++ test/fields/collections/Text/e2e.spec.ts
@@ -2,10 +2,10 @@ import type { Page } from '@playwright/test'
 import type { GeneratedTypes } from 'helpers/sdk/types.js'
 
 import { expect, test } from '@playwright/test'
-import { openListColumns } from 'helpers/e2e/openListColumns.js'
+import { openListColumns } from 'helpers/e2e/columns/openListColumns.js'
 import { upsertPreferences } from 'helpers/e2e/preferences.js'
-import { addListFilter } from 'helpers/e2e/tables/filters/addListFilter.js'
-import { toggleColumn } from 'helpers/e2e/toggleColumn.js'
+import { addListFilter } from 'helpers/e2e/filters/addListFilter.js'
+import { toggleColumn } from 'helpers/e2e/columns/toggleColumn.js'
 import path from 'path'
 import { wait } from 'payload/shared'
 import { fileURLToPath } from 'url'
