+++ test/group-by/e2e.spec.ts
@@ -4,11 +4,11 @@ import type { PayloadTestSDK } from 'helpers/sdk/index.js'
 import { expect, test } from '@playwright/test'
 import { devUser } from 'credentials.js'
 import { sortColumn } from 'helpers/e2e/columns/sortColumn.js'
+import { toggleColumn } from 'helpers/e2e/columns/toggleColumn.js'
 import { addListFilter } from 'helpers/e2e/filters/addListFilter.js'
 import { goToNextPage } from 'helpers/e2e/goToNextPage.js'
-import { addGroupBy, clearGroupBy, closeGroupBy, openGroupBy } from 'helpers/e2e/groupBy.js'
+import { addGroupBy, clearGroupBy, closeGroupBy, openGroupBy } from 'helpers/e2e/groupBy/index.js'
 import { deletePreferences } from 'helpers/e2e/preferences.js'
-import { toggleColumn } from 'helpers/e2e/toggleColumn.js'
 import { openNav } from 'helpers/e2e/toggleNav.js'
 import { reInitializeDB } from 'helpers/reInitializeDB.js'
 import * as path from 'path'
