+++ test/helpers/e2e/groupBy/index.js
@@ -1,0 +1,5 @@
+export { addGroupBy } from './addGroupBy.js'
+export { clearGroupBy } from './clearGroupBy.js'
+export { closeGroupBy } from './closeGroupBy.js'
+export { openGroupBy } from './openGroupBy.js'
+export { toggleGroupBy } from './toggleGroupBy.js'
