+++ test/i18n/e2e.spec.ts
@@ -13,7 +13,7 @@ import type { PayloadTestSDK } from '../helpers/sdk/index.js'
 
 import { ensureCompilationIsDone, initPageConsoleErrorCatch } from '../helpers.js'
 import { AdminUrlUtil } from '../helpers/adminUrlUtil.js'
-import { openListFilters } from '../helpers/e2e/tables/filters/openListFilters.js'
+import { openListFilters } from '../helpers/e2e/filters/openListFilters.js'
 import { initPayloadE2ENoConfig } from '../helpers/initPayloadE2ENoConfig.js'
 import { reInitializeDB } from '../helpers/reInitializeDB.js'
 import { TEST_TIMEOUT_LONG } from '../playwright.config.js'
