+++ test/uploads/e2e.spec.ts
@@ -2,9 +2,9 @@ import type { Page } from '@playwright/test'
 
 import { expect, test } from '@playwright/test'
 import { statSync } from 'fs'
-import { openListColumns } from 'helpers/e2e/openListColumns.js'
-import { openListFilters } from 'helpers/e2e/tables/filters/openListFilters.js'
-import { toggleColumn } from 'helpers/e2e/toggleColumn.js'
+import { openListColumns } from 'helpers/e2e/columns/openListColumns.js'
+import { toggleColumn } from 'helpers/e2e/columns/toggleColumn.js'
+import { openListFilters } from 'helpers/e2e/filters/openListFilters.js'
 import { openDocDrawer } from 'helpers/e2e/toggleDocDrawer.js'
 import path from 'path'
 import { wait } from 'payload/shared'
